back-end/requirements.txt
accelerate==1.11.0
annotated-types==0.7.0
anywidget==0.9.21
asttokens==3.0.0
bbox-visualizer==0.2.2
bidict==0.23.1
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
cython==3.2.1
decorator==5.2.1
defusedxml==0.7.1
einops==0.8.1
executing==2.2.1
fairscale==0.4.13
filelock==3.20.0
filetype==1.2.0
fonttools==4.60.1
fsspec==2025.9.0
ftfy==6.3.1
h11==0.16.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.7
ipython==9.7.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.8
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.2
jupyter-bbox-widget==0.6.0
jupyterlab-widgets==3.0.16
kiwisolver==1.4.9
markupsafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==2.2.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
open-clip-torch==3.2.0
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
packaging==25.0
pandas==2.3.3
parso==0.8.5
peft==0.18.0
pexpect==4.9.0
pi-heif==1.1.1
pillow==12.0.0
pillow-avif-plugin==1.5.2
polygraphy==0.49.26
prompt-toolkit==3.0.52
psutil==7.1.1
psygnal==0.15.0
ptyprocess==0.7.0
pure-eval==0.2.3
py-cpuinfo==9.0.0
pycocotools==2.0.10
pydantic==2.12.4
pydantic-core==2.41.5
pygments==2.19.2
pylabel==0.1.55
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.12.3
python-socketio==5.14.2
pyturbojpeg==1.8.2
pytz==2025.2
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rf100vl==1.1.0
rfdetr==1.3.0
roboflow==1.2.11
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
simple-websocket==1.1.0
six==1.17.0
stack-data==0.6.3
supervision==0.26.1
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.22
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
triton==3.5.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
ultralytics==8.3.156
ultralytics-thop==2.0.17
urllib3==2.5.0
wcwidth==0.2.14
websockets==12.0
werkzeug==3.1.3
widgetsnbextension==4.0.15
wsproto==1.2.0
